Extract casts to their own module
Simplifies ModelAttribute class; means that it doesn't add a 'cast' method to
classes that extend ModelAttribute.
